requirements.txt
h5py==2.10.0
numpy==1.17.3
opencv-python==4.1.1.26
pandas==0.25.3
Pillow==6.2.1
pytube==9.5.3
tensorflow==2.0.0